requirements.txt
aioice==0.9.0
aiortc==1.5.0
altair==5.1.2
attrs==23.1.0
av==10.0.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
dlib==19.24.2
dnspython==2.4.2
face-recognition==1.3.0
face_recognition_models==0.3.0
gitdb==4.0.10
GitPython==3.1.37
google-crc32c==1.5.0
ha-av==10.1.1
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
narwhals==1.28.0
numpy==1.26.0
opencv-python==4.11.0.86
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.4
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.1b0
pyee==11.0.0
Pygments==2.16.1
pylibsrtp==0.8.0
pymongo==4.5.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
streamlit==1.27.2
streamlit-webrtc==0.47.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
